pass error to ahoy to filter network errors from sentry

+++ Models/Queue.js
@@ -178,7 +178,7 @@ export class Queue {
     while (this.status === 'active' && concurrentJobs.length) {
 
       // Loop over jobs and process them concurrently.
-      const processingJobs = concurrentJobs.map( job => {
+      const processingJobs = concurrentJobs.map(job => {
         return this.processJob(job);
       });
 
@@ -276,10 +276,10 @@ export class Queue {
       // Grab concurrent job ids to reselect jobs as marking these jobs as active will remove
       // them from initial selection when write transaction exits.
       // See: https://stackoverflow.com/questions/47359368/does-realm-support-select-for-update-style-read-locking/47363356#comment81772710_47363356
-      const concurrentJobIds = jobsToMarkActive.map( job => job.id);
+      const concurrentJobIds = jobsToMarkActive.map(job => job.id);
 
       // Mark concurrent jobs as active
-      jobsToMarkActive = jobsToMarkActive.map( job => {
+      jobsToMarkActive = jobsToMarkActive.map(job => {
         job.active = true;
       });
 
@@ -324,8 +324,9 @@ export class Queue {
     this.worker.executeJobLifecycleCallback('onStart', jobName, jobId, jobPayload);
 
     try {
+      const executionResult = await this.worker.executeJob(job);
 
-      await this.worker.executeJob(job);
+      if (!executionResult.ok) throw new Error('Execution failure');
 
       // On successful job completion, remove job
       this.jobDB.delete(job);
@@ -348,7 +349,7 @@ export class Queue {
 
       // Log error
       if (!jobData.errors) {
-        jobData.errors = [ error.message ];
+        jobData.errors = [error.message];
       } else {
         jobData.errors.push(error.message);
       }
@@ -366,11 +367,11 @@ export class Queue {
       this.jobDB.save(job);
 
       // Execute job onFailure lifecycle callback.
-      this.worker.executeJobLifecycleCallback('onFailure', jobName, jobId, jobPayload);
+      this.worker.executeJobLifecycleCallback('onFailure', jobName, jobId, jobPayload, error);
 
       // If job has failed all attempts execute job onFailed and onComplete lifecycle callbacks.
       if (jobData.failedAttempts >= jobData.attempts) {
-        this.worker.executeJobLifecycleCallback('onFailed', jobName, jobId, jobPayload);
+        this.worker.executeJobLifecycleCallback('onFailed', jobName, jobId, jobPayload, error);
         this.worker.executeJobLifecycleCallback('onComplete', jobName, jobId, jobPayload);
       }
